requirements.txt
# requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.22
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
neo4j>=5.14.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.0
accelerate>=0.25.0